+++ teleop2servo/src/gamepad_teleop_node.cpp
@@ -193,8 +193,8 @@ bool GamepadTeleopNode::joy_in_use(
 
 bool GamepadTeleopNode::check_safety_procedure(const sensor_msgs::msg::Joy::SharedPtr & msg)
 {
-    const bool back_left = button_pressed(msg, Button::left_pad_left_click);
-    const bool back_right = button_pressed(msg, Button::left_pad_rigth_click);
+    const bool back_left = button_pressed(msg, Button::left_back_click);
+    const bool back_right = button_pressed(msg, Button::right_back_click);
     const bool b_pressed = rising_edge(msg, Button::b);
     const bool enable_sequence = back_left && back_right && b_pressed;
     {
@@ -310,10 +310,10 @@ void GamepadTeleopNode::create_cmd_joint(
     set_joint(1, Button::left_bumper, Button::right_bumper);
 
     if (!modifier) {
-        set_joint(2, Button::left_pad_top_click, Button::left_pad_down_click); // j3
+        set_joint(2, Button::left_pad_down_click, Button::left_pad_top_click); // j3
         set_joint(3, Button::left_pad_left_click, Button::left_pad_rigth_click);    // j4
     } else {
-        set_joint(4, Button::left_pad_top_click, Button::left_pad_down_click); // j5
+        set_joint(4, Button::left_pad_down_click, Button::left_pad_top_click); // j5
         set_joint(5, Button::left_pad_left_click, Button::left_pad_rigth_click);    // j6
     }