+++ teleop2servo/CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * [PR-01](https://github.com/AGH-CEAI/teleop2servo/pull/1) - Added joint conterol from the keyboard
 
 ### Changed
+
+* [PR-09](https://github.com/AGH-CEAI/teleop2servo/pull/9) - Small changes to gamepad key mapping
+
 ### Deprecated
 ### Removed
 ### Fixed